// src/types/enums.ts
export enum Role {
  ADMIN = "ADMIN",
  SAMPLING_HEAD = "SAMPLING_HEAD",
  SALES = "SALES",
}

export enum ItemStatus {
  REQUESTED = "REQUESTED",
  IN_SAMPLING = "IN_SAMPLING",
  SENT = "SENT",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
}

export enum Priority {
  LOW = "LOW",
  MEDIUM = "MEDIUM",
  HIGH = "HIGH",
}